fix: compilation errors and test fixes in Phase 1C Sprint 2 testing
- Fixed type mismatch: shiftAmount should be uint to prevent overflow (file_input.go:405)
- Removed unused 'io' import from file_output.go
- Fixed broken contains() helper function in tests using strings.Contains()
- Added time.Sleep in FileLocking test to avoid false positives
- Fixed unused variable in test (result -> _)
- Fixed docker-compose environment variable naming (NATS_URL -> FILE_INPUT_NATS_URL)
- Removed duplicate FILE_INPUT_SUBJECT in docker-compose-files.yml

All 83 unit/integration tests now pass successfully.

+++ src/pkg/io/file_input.go
@@ -35,15 +35,15 @@ type FileRetry struct {
 // FileConsumer monitors a directory for files and publishes them to NATS
 type FileConsumer struct {
 	// Configuration
-	dir                    string
-	pattern                string
-	pollInterval           time.Duration
-	archiveDir             string
-	errorDir               string
-	deleteAfterProcessing  bool
-	maxRetries             int
-	retryBackoffMs         int
-	archiveRetentionDays   int
+	dir                   string
+	pattern               string
+	pollInterval          time.Duration
+	archiveDir            string
+	errorDir              string
+	deleteAfterProcessing bool
+	maxRetries            int
+	retryBackoffMs        int
+	archiveRetentionDays  int
 
 	// Runtime
 	ctx            context.Context
@@ -401,11 +401,11 @@ func (f *FileConsumer) shouldRetry(filePath string) bool {
 	// Calculate backoff: exponential (1s, 2s, 4s, 8s, ...)
 	// Cap at reasonable value to prevent overflow: max 10 retries = 512x backoff = ~512 seconds
 	maxShift := uint(63) // Prevent overflow in bit shift
-	shiftAmount := retry.Attempts - 1
+	shiftAmount := uint(retry.Attempts - 1)
 	if shiftAmount > maxShift {
 		shiftAmount = maxShift
 	}
-	
+
 	backoffMs := f.retryBackoffMs * (1 << uint(shiftAmount))
 	backoffDuration := time.Duration(backoffMs) * time.Millisecond
